Moving Formula out of the init of codes.
